requirements.txt
streamlit == 23.3.1
plotly ==5.10.0
pandas == 1.4.3
numpy==1.23.1
folium ==0.14.0
matplotlib ==3.5.3
matplotlib-inline ==0.1.6
haversine==2.7.0
streamlit-folium ==0.15.0
pillow==9.2.0

